+++ crates/brr/src/lib.rs
@@ -24,7 +24,7 @@ pub const BYTES_PER_BRR_BLOCK: usize = 9;
 pub const BRR_HEADER_END_FLAG: u8 = 0x01;
 pub const BRR_HEADER_LOOP_FLAG: u8 = 0x02;
 
-#[derive(Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum BrrFilter {
     Filter0 = 0,
     Filter1 = 1,